+++ docs/maintaining/tagging_examples/docker_runner.py
@@ -3,4 +3,4 @@
 from tagging.utils.docker_runner import DockerRunner
 
 with DockerRunner("ubuntu") as container:
-    DockerRunner.run_simple_command(container, cmd="env", print_result=True)
+    DockerRunner.exec_cmd(container, cmd="env")