+++ tagging/taggers/sha.py
@@ -2,11 +2,8 @@
 # Distributed under the terms of the Modified BSD License.
 from docker.models.containers import Container
 
-from tagging.taggers.tagger_interface import TaggerInterface
 from tagging.utils.git_helper import GitHelper
 
 
-class SHATagger(TaggerInterface):
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        return GitHelper.commit_hash_tag()
+def commit_sha_tagger(container: Container) -> str:
+    return GitHelper.commit_hash_tag()