+++ tagging/taggers/tagger_interface.py
@@ -1,11 +1,7 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
+from collections.abc import Callable
+
 from docker.models.containers import Container
 
-
-class TaggerInterface:
-    """Common interface for all taggers"""
-
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        raise NotImplementedError
+TaggerInterface = Callable[[Container], str]