+++ tagging/utils/docker_runner.py
@@ -22,11 +22,10 @@ class DockerRunner:
         self.docker_client: docker.DockerClient = docker_client
 
     def __enter__(self) -> Container:
-        LOGGER.info(f"Creating container for image {self.image_name} ...")
+        LOGGER.info(f"Creating a container for the image: {self.image_name} ...")
+        default_kwargs = {"detach": True, "tty": True}
         self.container = self.docker_client.containers.run(
-            image=self.image_name,
-            command=self.command,
-            detach=True,
+            image=self.image_name, command=self.command, **default_kwargs
         )
         LOGGER.info(f"Container {self.container.name} created")
         return self.container
@@ -43,14 +42,11 @@ class DockerRunner:
         LOGGER.info(f"Container {self.container.name} removed")
 
     @staticmethod
-    def run_simple_command(
-        container: Container, cmd: str, print_result: bool = True
-    ) -> str:
-        LOGGER.info(f"Running cmd: '{cmd}' on container: {container}")
-        out = container.exec_run(cmd)
-        result = out.output.decode().rstrip()
-        assert isinstance(result, str)
-        if print_result:
-            LOGGER.info(f"Command result: {result}")
-        assert out.exit_code == 0, f"Command: {cmd} failed"
-        return result
+    def exec_cmd(container: Container, cmd: str) -> str:
+        LOGGER.info(f"Running cmd: `{cmd}` on container: {container.name}")
+        exec_result = container.exec_run(cmd)
+        output = exec_result.output.decode().rstrip()
+        assert isinstance(output, str)
+        LOGGER.debug(f"Command output: {output}")
+        assert exec_result.exit_code == 0, f"Command: `{cmd}` failed"
+        return output