+++ tagging/utils/quoted_output.py
@@ -1,12 +1,14 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
+import textwrap
+
 from docker.models.containers import Container
 
 from tagging.utils.docker_runner import DockerRunner
 
 
 def quoted_output(container: Container, cmd: str) -> str:
-    cmd_output = DockerRunner.run_simple_command(container, cmd, print_result=False)
+    cmd_output = DockerRunner.exec_cmd(container, cmd)
     # For example, `mamba info` adds redundant empty lines
     cmd_output = cmd_output.strip("\n")
     # For example, R packages list contains trailing backspaces
@@ -14,9 +16,11 @@ def quoted_output(container: Container, cmd: str) -> str:
 
     assert cmd_output, f"Command `{cmd}` returned empty output"
 
-    return f"""\
-`{cmd}`:
+    return textwrap.dedent(
+        f"""\
+        `{cmd}`:
 
-```text
-{cmd_output}
-```"""
+        ```text
+        {{output}}
+        ```"""
+    ).format(output=cmd_output)